requirements.txt
Flask
flask-cors
gunicorn
python-telegram-bot